% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{add_missing_rows}
\alias{add_missing_rows}
\title{Function to add missing rows to a matrix}
\usage{
add_missing_rows(data, rows, sorted = TRUE)
}
\arguments{
\item{data}{matrix, rows = genes, cols = samples}

\item{rows}{vector of rownames to add}

\item{sorted}{logical, whether to sort the rows alphabetically}
}
\value{
data matrix, rows = genes, cols = samples
}
\description{
Function to add missing rows to a matrix
}
