% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_functions.R
\name{align_signature}
\alias{align_signature}
\title{Align a signature by average correlation with the derivation values}
\usage{
align_signature(sig, data, by = "mean")
}
\arguments{
\item{sig}{signature vector}

\item{data}{data matrix}

\item{by}{method to align by (default is "mean")}
}
\value{
aligned signature
}
\description{
Align a signature by average correlation with the derivation values
}
