% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_functions.R
\name{calculate_correlations}
\alias{calculate_correlations}
\title{Calculate Correlations}
\usage{
calculate_correlations(
  dds,
  condition,
  normalize = "mor",
  method = "spearman",
  ...
)
}
\arguments{
\item{dds}{DESeq2 object}

\item{condition}{Column of interest}

\item{normalize}{Normalization method to use}

\item{method}{Correlation method to use}

\item{...}{Additional arguments to pass to cor.test}
}
\value{
Correlation results for each gene
}
\description{
Calculate Correlations
}
