% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_functions.R
\name{check_input_data}
\alias{check_input_data}
\title{Function to check input data}
\usage{
check_input_data(data, required_columns)
}
\arguments{
\item{data}{Data frame containing the data to be checked}

\item{required_columns}{Vector of required column names}
}
\value{
Logical value indicating whether the data is valid
}
\description{
Function to check input data
}
