% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{clustering_functions
Prepare a counts matrix for NMF}
\alias{clustering_functions
Prepare a counts matrix for NMF}
\alias{prep_nmf_data}
\title{Clustering Functions}
\usage{
prep_nmf_data(counts_matr)
}
\arguments{
\item{counts_matr}{A matrix with rows as genes and columns as samples.}
}
\value{
A combined matrix with positive and negative values separated.
}
\description{
Functions for preparing data and performing clustering using NMF, k-means, and hierarchical clustering.
}
