% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{clustering_functions}
\alias{clustering_functions}
\title{Clustering Functions}
\description{
Functions for preparing data and performing clustering using NMF, k-means, and hierarchical clustering.
}
\details{
This script contains functions to prepare a counts matrix for NMF, estimate the rank of a matrix using NMF, plot the results of NMF, estimate k-means clustering using the elbow method, and estimate hierarchical clustering.
}
