% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{color_mapping}
\alias{color_mapping}
\title{Color Mapping}
\usage{
color_mapping(df, custom_colors = NULL)
}
\arguments{
\item{df}{Data frame of annotations}

\item{custom_colors}{List of custom colors to use with names of variables}
}
\value{
List of annotation color maps
}
\description{
Function to make annotation color maps for continuous variables and categorical variables
}
