% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_functions.R
\name{compare_one_to_many}
\alias{compare_one_to_many}
\title{Compare one column to many columns}
\usage{
compare_one_to_many(
  df,
  col,
  cols,
  outdir,
  plot = TRUE,
  method = "spearman",
  ...
)
}
\arguments{
\item{df}{data frame}

\item{col}{column to compare}

\item{cols}{columns to compare to}

\item{outdir}{output directory}

\item{plot}{logical, whether to plot the results}

\item{method}{correlation method (default is "spearman")}

\item{...}{additional arguments to pass to stats functions}
}
\value{
list of stats results and plots
}
\description{
Compare one column to many columns
}
