% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_functions.R
\name{compare_results}
\alias{compare_results}
\title{Compare DESeq Results}
\usage{
compare_results(
  res1,
  res2,
  metric = "log2FoldChange",
  by = "rowname",
  suffix = c("_1", "_2")
)
}
\arguments{
\item{res1}{Results of differential expression analysis 1.}

\item{res2}{Results of differential expression analysis 2.
data frame of comparison results}
}
\description{
Function to compare different DESeq results.
}
