% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converting_functions.R
\name{converting_functions}
\alias{converting_functions}
\title{Converting Functions}
\description{
This script contains functions for converting gene IDs and normalizing data.
}
\details{
The functions include mapping gene IDs using `AnnotationDbi`, converting gene IDs using `biomaRt`, detecting gene ID types, and normalizing data.
}
