% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_cores}
\alias{detect_cores}
\title{Detect the number of cores on a machine}
\usage{
detect_cores()
}
\value{
cores integer, number of cores on the machine
}
\description{
Detect the number of cores on a machine
}
