% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converting_functions.R
\name{detect_gene_id_type}
\alias{detect_gene_id_type}
\title{Detect Gene ID Type}
\usage{
detect_gene_id_type(geneList, strip = TRUE)
}
\arguments{
\item{geneList}{list or vector of gene IDs.}

\item{strip}{logical, whether to strip version numbers from gene IDs.}
}
\value{
character, type of gene ID.
}
\description{
Function to detect gene ID type based on the first gene ID.
}
