% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_threads}
\alias{detect_threads}
\title{Detect the number of threads on a machine}
\usage{
detect_threads()
}
\value{
threads integer, number of threads on the machine
}
\description{
Detect the number of threads on a machine
}
