% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering_functions.R
\name{detect_wbc}
\alias{detect_wbc}
\title{Detect WBC contamination in platelet RNA-seq}
\usage{
detect_wbc(dds, outpath, group = NULL, normalize = "vst")
}
\arguments{
\item{dds}{DESeq2 object}

\item{outpath}{Path to output directory}

\item{group}{Column of interest}

\item{normalize}{Normalization method}
}
\value{
DESeq2 object after detecting WBC contamination
}
\description{
Detect WBC contamination in platelet RNA-seq
}
