% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop_na_cols}
\alias{drop_na_cols}
\title{Drop columns with NA values based on a threshold}
\usage{
drop_na_cols(data, percent_allowed_missing)
}
\arguments{
\item{data}{data frame}

\item{percent_allowed_missing}{numeric, threshold for allowed missing values}
}
\value{
data frame with columns dropped
}
\description{
Drop columns with NA values based on a threshold
}
