% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop_na_rows}
\alias{drop_na_rows}
\title{Drop rows with NA values based on a threshold}
\usage{
drop_na_rows(data, percent_allowed_missing)
}
\arguments{
\item{data}{data frame}

\item{percent_allowed_missing}{numeric, threshold for allowed missing values}
}
\value{
data frame with rows dropped
}
\description{
Drop rows with NA values based on a threshold
}
