% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_functions.R
\name{eigen_nmf}
\alias{eigen_nmf}
\title{Calculate eigengenes by non-negative matrix factorization (WIP)}
\usage{
eigen_nmf(
  df,
  outdir,
  pcs = 1,
  align = FALSE,
  method = "brunet",
  seed = 420,
  nrun = 1
)
}
\arguments{
\item{df}{data frame [samples x genes]}

\item{outdir}{output directory}

\item{pcs}{number of principal components to return}

\item{align}{logical, align eigengenes by average expression}

\item{method}{NMF algorithm method}

\item{seed}{random seed for reproducibility}

\item{nrun}{number of runs}
}
\value{
dataframe with eigengenes
}
\description{
Calculate eigengenes by non-negative matrix factorization
}
