% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_functions.R
\name{eigen_pca}
\alias{eigen_pca}
\title{Calculate eigengenes by principal component analysis}
\usage{
eigen_pca(df, outdir, pcs = 1, align = TRUE, ...)
}
\arguments{
\item{df}{data frame [samples x genes]}

\item{outdir}{output directory}

\item{pcs}{number of principal components to return}

\item{align}{logical, align eigengenes by average expression}

\item{...}{additional arguments to pass to stats functions}
}
\value{
dataframe with eigengenes
}
\description{
Calculate eigengenes by principal component analysis
}
