% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_functions.R
\name{eigen_reg_svd}
\alias{eigen_reg_svd}
\title{Calculate eigengenes by regularized singular value decomposition (WIP)}
\usage{
eigen_reg_svd(
  M,
  samples = nrow(M),
  vectors = 1:3,
  tau = 1,
  lap = FALSE,
  method = "euclidian",
  verbose = FALSE
)
}
\arguments{
\item{M}{data matrix [samples x genes]}

\item{samples}{number of samples}

\item{vectors}{number of eigenvectors to return}

\item{tau}{tau value for distance matrix}

\item{lap}{logical, use laplacian distance}

\item{method}{distance method}

\item{verbose}{logical, print verbose output}
}
\value{
dataframe with eigengenes
}
\description{
Calculate eigengenes by regularized singular value decomposition
}
