% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converting_functions.R
\name{fetch_gene_coordinates}
\alias{fetch_gene_coordinates}
\title{Fetch Gene Coordinates from Bioconductor}
\usage{
fetch_gene_coordinates(
  genes,
  key = "SYMBOL",
  genome_build = "hg38",
  filter_chromosomes = "^chr([1-9]|1[0-9]|2[0-2]|X|Y)$"
)
}
\arguments{
\item{genes}{Character vector of gene symbols}

\item{genome_build}{Genome build (default: "hg38")}
}
\value{
Data frame with gene, chr, start, end columns
}
\description{
Function to fetch gene coordinates from Bioconductor based on gene symbols and genome build.
}
