% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering_functions.R
\name{filter_edgeR}
\alias{filter_edgeR}
\title{Filter a DESeq2 object by expression and make plots of the filtering}
\usage{
filter_edgeR(dds, outpath, group = NULL, min.count = 10, min.prop = 0.5, ...)
}
\arguments{
\item{dds}{DESeq2 object}

\item{outpath}{Path to output directory}

\item{group}{Column of interest}

\item{min.count}{Minimum mean expression to keep}

\item{min.prop}{Minimum proportion of samples to keep}

\item{...}{Additional arguments to pass to FilterByExpr}
}
\value{
DESeq2 object after filtering
}
\description{
Filter a DESeq2 object by expression and make plots of the filtering
}
