% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_functions.R
\name{gene_wilcox_test}
\alias{gene_wilcox_test}
\title{Wilcoxon Ranked Sum testing on genes in two summarized experiments}
\usage{
gene_wilcox_test(dds, outpath, condition, pCutoff = 0.05, FCcutoff = 0.5, ...)
}
\arguments{
\item{dds}{DESeq2 object}

\item{outpath}{Output path for results}

\item{condition}{Column of interest}

\item{pCutoff}{p-value cutoff for volcano plot}

\item{FCcutoff}{Fold change cutoff for volcano plot}

\item{...}{Additional arguments to pass to wilcox.test}
}
\value{
Data frame of p-values and adjusted p-values
}
\description{
Wilcoxon Ranked Sum testing on genes in two summarized experiments
}
