% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{general_functions}
\alias{general_functions}
\alias{summarize_experiment}
\title{General functions}
\usage{
summarize_experiment(
  results,
  logFC_column = "log2FoldChange",
  pvalue_column = "pvalue",
  padj_column = "padj",
  pvalue_cutoffs = c(0.01, 0.05, 0.1),
  padj_cutoffs = c(0.05, 0.1, 0.2),
  logFC_cutoff = 0
)
}
\arguments{
\item{results}{results of differential expression analysis}

\item{logFC_column}{column to use for logFC}

\item{pvalue_column}{column to use for pvalue}

\item{padj_column}{column to use for padj}

\item{pvalue_cutoffs}{list of pvalue cutoffs to use}

\item{padj_cutoffs}{list of padj cutoffs to use}

\item{logFC_cutoff}{logFC cutoff to use}
}
\value{
summary of results
}
\description{
General functions for data manipulation and analysis
}
