% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_functions.R
\name{get_eGFR}
\alias{get_eGFR}
\title{Function to calculate eGFR}
\usage{
get_eGFR(
  age,
  sex,
  creatinine,
  sex_levels = c("Male", "Female"),
  creatinine_units = c("mg/dL", "umol/L")
)
}
\arguments{
\item{age}{Age of the individual}

\item{sex}{vector of sex}

\item{creatinine}{Serum creatinine level (mg/dL)}
}
\value{
Estimated glomerular filtration rate (eGFR) in mL/min/1.73m²
}
\description{
Function to calculate estimated glomerular filtration rate (eGFR) according to the CKD-EPI (2021) equation.
}
