% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{get_events_breakdown}
\alias{get_events_breakdown}
\title{Function to get the number of samples in each group}
\usage{
get_events_breakdown(
  metadata,
  id = "PATNUM",
  events_term = "C_",
  subset = NULL
)
}
\arguments{
\item{metadata}{dataframe, metadata}

\item{id}{character, column name of the sample ID}

\item{events_term}{character, prefix of the events columns}

\item{subset}{character, column name of the grouping variable}
}
\value{
dataframe with the number of samples in each group
}
\description{
Function to get the number of samples in each group
}
