% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{getGenes}
\alias{getGenes}
\title{Function to get genes from a results table}
\usage{
getGenes(
  results,
  pval = 0.05,
  metric = 0,
  name_col = "rownames",
  pval_col = "padj",
  metric_col = "log2FoldChange"
)
}
\arguments{
\item{results}{results table}

\item{pval}{p-value cutoff}

\item{pval_col}{p-value column name}

\item{log2fc}{log2 fold change cutoff}

\item{gene_col}{gene column name}

\item{log2fc_col}{log2 fold change column name}
}
\value{
dataframe of genes separated by up and down
}
\description{
Function to get genes from a results table
}
