% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_functions.R
\name{gsea_analysis}
\alias{gsea_analysis}
\title{Run a GSEA analysis}
\usage{
gsea_analysis(
  geneList,
  outpath,
  keyType = NULL,
  msigdb_category = "H",
  ontology = "ALL"
)
}
\arguments{
\item{geneList}{List of genes to run enrichment on.}

\item{outpath}{Path to save results.}

\item{keyType}{Key type for gene list.}

\item{msigdb_category}{MSigDB category (default is "H").}

\item{ontology}{Ontology to use (default is ALL).}
}
\value{
Enrichment results for each level of column of interest.
}
\description{
Run a GSEA analysis
}
