% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_functions.R
\name{handle_deseq_output}
\alias{handle_deseq_output}
\title{Handle DESeq Output}
\usage{
handle_deseq_output(
  dds,
  res,
  resLFC,
  outpath,
  name,
  pvalue,
  pCutoff,
  fcCutoff,
  contrast
)
}
\arguments{
\item{dds}{DESeq2 object}

\item{res}{Results of differential expression analysis}

\item{resLFC}{Results with log fold change shrinkage}

\item{outpath}{Output path for results}

\item{name}{Name of the analysis}

\item{pvalue}{p-value column name}

\item{pCutoff}{p-value cutoff for volcano plot}

\item{fcCutoff}{Fold change cutoff for volcano plot}

\item{contrast}{Contrast used in the analysis}
}
\value{
None
}
\description{
Function to handle DESeq output and generate plots and results
}
\keyword{internal}
