% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{hazard_ratios_table}
\alias{hazard_ratios_table}
\title{Hazard Ratios Table}
\usage{
hazard_ratios_table(
  df,
  condition,
  censors,
  controls = NULL,
  per_sd = FALSE,
  ovr = FALSE,
  subgroups = NULL,
  censor_prefix = "censor_",
  time_prefix = "time_to_",
  verbose = FALSE
)
}
\arguments{
\item{df}{data.frame, data to make survival curves from.}

\item{condition}{character, condition(s) to make survival curves for. Can be a single condition or vector of conditions.}

\item{censors}{character, censors to make survival curves for.}

\item{controls}{character, controls to include in the model. Default is NULL.}

\item{per_sd}{logical, whether to standardize the condition. Default is FALSE.}

\item{ovr}{logical, whether to do one vs rest. Default is FALSE.}

\item{subgroups}{character vector, variable names for subgroup analyses. Default is NULL.}

\item{censor_prefix}{character, prefix for censor columns. Default is 'censor_'.}

\item{time_prefix}{character, prefix for time columns. Default is 'time_to_'.}

\item{verbose}{logical, whether to print verbose messages. Default is FALSE.}
}
\value{
data.frame, hazard ratios table.
}
\description{
Make HR tables for one or more conditions and censors using univariate analyses.
}
