% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{hazards.internal}
\alias{hazards.internal}
\title{Internal Hazard Ratios Helper}
\usage{
hazards.internal(cs, df, condition, controls)
}
\arguments{
\item{cs}{list, list of censors and time variables.}
}
\description{
Internal function to calculate hazard ratios.
}
