% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{hclust_estimator}
\alias{hclust_estimator}
\title{Estimate hierarchical clustering of a matrix}
\usage{
hclust_estimator(counts_matr, clusters = 2:8, outfile, ...)
}
\arguments{
\item{counts_matr}{A matrix with rows as genes and columns as samples.}

\item{clusters}{A vector of clusters to test. Default is 2:8.}

\item{outfile}{A file to save the output to.}

\item{...}{Other arguments to pass to hclust.}
}
\value{
A list of hierarchical clustering results.
}
\description{
Estimate hierarchical clustering of a matrix
}
