% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{hypergeometric_scoring}
\alias{hypergeometric_scoring}
\title{Hypergeometric Scoring}
\usage{
hypergeometric_scoring(gse, method = "fisher", ...)
}
\arguments{
\item{gse}{gse object from hypergeometric enrichment analysis in clusterProfiler}

\item{method}{character, method to use ('fisher' or 'chisq')}

\item{...}{other arguments to pass to the test function}
}
\value{
gse object with odds ratio and p-value added
}
\description{
Add odds ratio and p-value to a gse object from hypergeometric enrichment analysis
}
