% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_packages_from_script}
\alias{install_packages_from_script}
\title{Install all required packages from a script}
\usage{
install_packages_from_script(script)
}
\arguments{
\item{script}{character, path to script}
}
\value{
none
}
\description{
Install all required packages from a script
}
