% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{list_of_lists_to_df}
\alias{list_of_lists_to_df}
\title{Function to turn list of lists into a dataframe}
\usage{
list_of_lists_to_df(list)
}
\arguments{
\item{list}{list of lists}
}
\value{
dataframe
}
\description{
Function to turn list of lists into a dataframe
}
