% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_to_json}
\alias{list_to_json}
\title{Generate a JSON file from a list}
\usage{
list_to_json(list, filename)
}
\arguments{
\item{list}{list, list to convert to JSON}

\item{filename}{character, name of the file to write}
}
\value{
none
}
\description{
Generate a JSON file from a list
}
