% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenomics_functions.R
\name{make_compsite_coding_csv}
\alias{make_compsite_coding_csv}
\title{Make Composite Coding CSV}
\usage{
make_compsite_coding_csv(lol, composite_name)
}
\arguments{
\item{lol}{List of lists containing the codes.}

\item{composite_name}{Name of the composite variable.}
}
\value{
A data frame with the composite variable.
}
\description{
Function to take any codes and make a composite variable CSV.
}
