% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{make_dds}
\alias{make_dds}
\title{Function to make a DESeqDataSet object}
\usage{
make_dds(countsMatr, colData, design = ~1)
}
\arguments{
\item{countsMatr}{matrix of counts, rows = genes, cols = samples}

\item{colData}{data.frame of sample metadata, rows = samples}

\item{design}{formula for the design matrix}
}
\value{
DESeqDataSet object
}
\description{
Function to make a DESeqDataSet object
}
