% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenomics_functions.R
\name{make_phenotypes}
\alias{make_phenotypes}
\title{Make Phenotypes}
\usage{
make_phenotypes(hesin, encoding)
}
\arguments{
\item{hesin}{Data frame of hesin data pivoted longer [dnx_hesin_id, Participant ID, field, value].}

\item{encoding}{Data frame of composite encoding [key, field, value].}
}
\value{
A data frame with composite encoding.
}
\description{
Function to make a phenotype from the composite encoding.
}
