% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{make_se}
\alias{make_se}
\title{Function to make a SummarizedExperiment object}
\usage{
make_se(countsMatr, colData)
}
\arguments{
\item{countsMatr}{matrix of counts, rows = genes, cols = samples}

\item{colData}{data.frame of sample metadata, rows = samples}
}
\value{
SummarizedExperiment object
}
\description{
Function to make a SummarizedExperiment object
}
