% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_functions.R
\name{maybe_align}
\alias{maybe_align}
\title{Maybe align signature scores}
\usage{
maybe_align(mat, align, data)
}
\arguments{
\item{mat}{Matrix of scores to align}

\item{align}{Logical, whether to perform alignment}

\item{data}{Original data frame used for alignment reference}
}
\value{
Aligned matrix if align=TRUE, otherwise original matrix
}
\description{
Internal helper to optionally align signature scores with data.
}
\keyword{internal}
