% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{min_max_norm}
\alias{min_max_norm}
\title{Min-Max Normalize a Vector}
\usage{
min_max_norm(x)
}
\arguments{
\item{x}{numeric, vector to min-max normalize}
}
\value{
numeric, min-max normalized vector
}
\description{
Min-Max Normalize a Vector
}
