% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mpa_geneset}
\alias{mpa_geneset}
\title{MPA Geneset
 
A dataframe containing the MPA geneset}
\format{
A data frame with 42 rows and 2 variables:
\describe{
  \item{gs_name}{Geneset name}
  \item{gene_symbol}{Gene symbol}
}
}
\source{
\url{TBD}
}
\usage{
mpa_geneset
}
\description{
MPA Geneset
 
A dataframe containing the MPA geneset
}
\keyword{datasets}
