% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{nmf_estimator}
\alias{nmf_estimator}
\title{Estimate the rank of a matrix using NMF}
\usage{
nmf_estimator(counts_matr, outfile, ranks = 2:8, runs = 30, options = "v", ...)
}
\arguments{
\item{counts_matr}{A matrix with rows as genes and columns as samples.}

\item{outfile}{A file to save the output to.}

\item{ranks}{A vector of ranks to test. Default is 2:8.}

\item{runs}{Number of runs to test. Default is 30.}

\item{options}{Options for NMF. Default is 'v'.}

\item{...}{Other arguments to pass to NMF.}
}
\value{
A list of NMF and random NMF results.
}
\description{
Estimate the rank of a matrix using NMF
}
