% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{nmf_plotter}
\alias{nmf_plotter}
\title{Plot the results of NMF estimation}
\usage{
nmf_plotter(nmf_out)
}
\arguments{
\item{nmf_out}{Output from \code{nmf_estimator}.}
}
\value{
A ggplot object.
}
\description{
Plot the results of NMF estimation
}
