% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_functions.R
\name{olink_analysis}
\alias{olink_analysis}
\title{Run differential expression and pathway analysis}
\usage{
olink_analysis(
  data,
  conditions,
  outdir,
  de_fxn = OlinkAnalyze::olink_wilcox,
  gsea_fxn = OlinkAnalyze::olink_pathway_enrichment,
  de_args = list(),
  volcano_args = list(),
  gsea_args = list()
)
}
\arguments{
\item{data}{Data frame of Olink data}

\item{conditions}{Condition column}

\item{outdir}{Output directory}

\item{de_fxn}{Differential expression function}

\item{gsea_fxn}{Pathway analysis function}

\item{de_args}{Arguments to pass to differential expression function}

\item{volcano_args}{Arguments to pass to volcano plot function}

\item{gsea_args}{Arguments to pass to pathway analysis function}
}
\value{
List of differential expression and pathway analysis results
}
\description{
Run differential expression and pathway analysis on Olink data
}
