% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_functions.R
\name{olink_count_table}
\alias{olink_count_table}
\title{Convert Olink data to count table}
\usage{
olink_count_table(data, sampleID = "SampleID", assay = "Assay", value = "NPX")
}
\arguments{
\item{data}{Data frame of Olink data}

\item{sampleID}{Sample ID column}

\item{assay}{Assay column}

\item{value}{Value column}
}
\value{
Count table
}
\description{
Convert Olink data to a count table
}
