% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_functions.R
\name{olink_filtering}
\alias{olink_filtering}
\title{Perform Olink filtering}
\usage{
olink_filtering(
  data,
  outdir,
  pca_args = list(),
  umap_args = list(),
  lod_args = list()
)
}
\arguments{
\item{data}{Data frame of Olink data}

\item{outdir}{Output directory}

\item{pca_args}{List of arguments to pass to PCA function}

\item{umap_args}{List of arguments to pass to UMAP function}

\item{lod_args}{List of arguments to pass to LOD function}
}
\value{
List containing filtered data, count table, and outliers
}
\description{
Perform PCA, UMAP, and LOD filtering on Olink data
}
