% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_functions.R
\name{olink_info}
\alias{olink_info}
\title{Get Olink sample info}
\usage{
olink_info(data)
}
\arguments{
\item{data}{Data frame of Olink data}
}
\value{
List containing sample IDs, plates, proteins, and panels
}
\description{
Get sample IDs, plates, proteins, and panels from Olink data
}
