% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_functions.R
\name{olink_lod_qc}
\alias{olink_lod_qc}
\title{Test level of detection (LOD) for each protein}
\usage{
olink_lod_qc(data, outdir, plot = TRUE)
}
\arguments{
\item{data}{Data frame of Olink data}

\item{outdir}{Output directory}

\item{plot}{Boolean to indicate if plots should be generated}
}
\value{
List containing LOD results and outliers
}
\description{
Test level of detection (LOD) for each protein
}
