% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_functions.R
\name{olink_odds_ratios}
\alias{olink_odds_ratios}
\title{Calculate odds ratios for Olink data}
\usage{
olink_odds_ratios(data, proteins, events, adjustments = NULL)
}
\arguments{
\item{data}{Data frame of Olink data}

\item{proteins}{Vector of proteins to calculate odds ratios for}

\item{events}{Vector of events to calculate odds ratios for}

\item{adjustments}{Vector of variables to adjust for}
}
\value{
List of odds ratios for each event
}
\description{
Calculate odds ratios for Olink data
}
