% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_functions.R
\name{olink_umap_outliers}
\alias{olink_umap_outliers}
\title{Get Olink UMAP outliers and plot}
\usage{
olink_umap_outliers(
  data,
  outpath,
  outlierDefX = 2.5,
  outlierDefY = 4,
  byPanel = TRUE,
  ...
)
}
\arguments{
\item{data}{Data frame of Olink data}

\item{outpath}{Output directory}

\item{outlierDefX}{Outlier definition for x axis (default: 2.5)}

\item{outlierDefY}{Outlier definition for y axis (default: 4)}

\item{byPanel}{Boolean to plot by panel (default: TRUE)}

\item{...}{Additional arguments to pass to olink_umap_plot}
}
\value{
List containing UMAP results and outliers
}
\description{
Get UMAP outliers from Olink data
}
