% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{one_hot_encode_ovr}
\alias{one_hot_encode_ovr}
\title{Function to one hot encode dataframe column}
\usage{
one_hot_encode_ovr(df, column, binary = TRUE)
}
\arguments{
\item{df}{dataframe}

\item{column}{character, column name}

\item{binary}{logical, whether to make the column binary}
}
\value{
dataframe with one hot encoded column
}
\description{
Function to one hot encode dataframe column
}
