% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{pairwise_combos}
\alias{pairwise_combos}
\title{Function to get make pairwise combinations}
\usage{
pairwise_combos(vec)
}
\arguments{
\item{vec}{vector of values}
}
\value{
list of pairwise combinations
}
\description{
Function to get make pairwise combinations
}
