% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenomics_functions.R
\name{phenomics_functions
Make Composite Coding CSV}
\alias{phenomics_functions
Make Composite Coding CSV}
\alias{make_composite_coding}
\title{Phenomics Functions}
\usage{
make_composite_coding(lol, composite_name)
}
\arguments{
\item{lol}{List of lists containing the codes}

\item{composite_name}{Name of the composite variable}
}
\value{
Data frame with composite variable mapping
}
\description{
Functions for EHR / Biobank data processing.
}
