% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_clinical_factors_heatmap}
\alias{plot_clinical_factors_heatmap}
\title{Plot Clinical Factors Heatmap}
\usage{
plot_clinical_factors_heatmap(
  cohort_tab,
  name = "Statistic",
  na_col = "#e2dede",
  cluster_rows = FALSE,
  cluster_columns = FALSE,
  show_row_names = TRUE,
  show_column_names = TRUE,
  row_title = "Clinical Factors",
  row_title_gp = gpar(fontsize = 12),
  row_names_gp = gpar(fontsize = 16),
  column_names_gp = gpar(fontsize = 16),
  column_names_rot = 0,
  col = colorRamp2(c(0, 1, 5), c("#acacf7", "white", "red")),
  border = TRUE,
  rect_gp = gpar(col = "black", lwd = 1),
  width = unit(0.6 * nrow(cohort_tab), "cm"),
  height = unit(1.4 * nrow(cohort_tab), "cm"),
  ...
)
}
\arguments{
\item{cohort_tab}{Data frame to plot}

\item{name}{Name of the heatmap}

\item{na_col}{Color for NA values}

\item{cluster_rows}{Logical, whether to cluster rows}

\item{cluster_columns}{Logical, whether to cluster columns}

\item{show_row_names}{Logical, whether to show row names}

\item{show_column_names}{Logical, whether to show column names}

\item{row_title}{Title for the rows}

\item{row_title_gp}{Graphical parameters for row title}

\item{row_names_gp}{Graphical parameters for row names}

\item{column_names_gp}{Graphical parameters for column names}

\item{column_names_rot}{Rotation for column names}

\item{col}{Color scale for the heatmap}

\item{border}{Logical, whether to show borders}

\item{rect_gp}{Graphical parameters for rectangles}

\item{width}{Width of the heatmap}

\item{height}{Height of the heatmap}

\item{...}{Additional arguments to pass to ComplexHeatmap}
}
\value{
ComplexHeatmap object
}
\description{
Function to plot a clinical factors heatmap
}
