% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_correlation_matrix}
\alias{plot_correlation_matrix}
\title{Plot Correlation Matrix}
\usage{
plot_correlation_matrix(
  cor_mat,
  title = "",
  xlab = "",
  ylab = "",
  x_order = NULL,
  y_order = NULL,
  ...
)
}
\arguments{
\item{cor_mat}{Data frame, correlation matrix}

\item{title}{Title of the plot}

\item{xlab}{X-axis label}

\item{ylab}{Y-axis label}

\item{x_order}{Order of x-axis variables}

\item{y_order}{Order of y-axis variables}

\item{...}{Additional arguments to pass to ggplot}
}
\value{
ggplot object
}
\description{
Function to plot the correlation matrix
}
