% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_enrichment}
\alias{plot_enrichment}
\title{Plot Enrichment Terms}
\usage{
plot_enrichment(
  gse,
  title = "Enrichment Plot",
  terms2plot = NULL,
  genes2plot = NULL,
  padjCutoff = 0.2,
  max_terms = 20,
  wrap_width = 40
)
}
\arguments{
\item{gse}{gse object}

\item{title}{Title of plot}

\item{terms2plot}{List of terms to plot (optional)}

\item{genes2plot}{List of genes to plot (optional)}

\item{max_terms}{Maximum number of terms to plot (default: 20)}

\item{wrap_width}{Width for wrapping term descriptions (default: 40)}

\item{qvalueCutoff}{Q-value cutoff (default: 0.2)}
}
\value{
ggplot object of enrichment terms
}
\description{
Function to plot enrichment from gse object
}
