% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_ggplot_table}
\alias{plot_ggplot_table}
\title{Plot ggplot Table}
\usage{
plot_ggplot_table(df, columns, axis, size = 5, ...)
}
\arguments{
\item{df}{Data frame to plot}

\item{columns}{Columns to plot}

\item{axis}{Axis to plot}

\item{size}{Size of text}

\item{...}{Additional arguments to pass to ggplot}
}
\value{
ggplot object of table
}
\description{
Function to take a dataframe and make it into a ggplot table with each column being its own ggplot object
}
