% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot Gene Heatmap}
\usage{
plot_heatmap(dds, genes = NULL, annotations = NULL, normalize = "vst", ...)
}
\arguments{
\item{dds}{DESeq2 object}

\item{annotations}{List of annotations to plot}

\item{normalize}{Type of normalization to use}

\item{...}{Additional arguments to pass to ComplexHeatmap}

\item{title}{Title of plot}
}
\value{
ComplexHeatmap object of gene heatmap
}
\description{
Function to plot complex heatmap from DESeq2 object
}
