% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_library_depth}
\alias{plot_library_depth}
\title{Plotting Functions}
\usage{
plot_library_depth(dds, title, bins = 30)
}
\arguments{
\item{dds}{DESeq2 object}

\item{title}{Title of plot}

\item{bins}{Number of bins for histogram}
}
\value{
ggplot object of library depth
}
\description{
Functions for plotting data.

Function to plot library depth of summarized experiment object
}
\author{
Matthew Muller
}
