% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_library_size}
\alias{plot_library_size}
\title{Plot Library Size}
\usage{
plot_library_size(dds, title, bins = 10)
}
\arguments{
\item{dds}{DESeq2 object}

\item{title}{Title of plot}

\item{bins}{Number of bins for histogram (default: 10)}
}
\value{
ggplot object of library size histogram
}
\description{
Function to plot library size of summarized experiment object
}
