% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_odds_volcano}
\alias{plot_odds_volcano}
\title{Plot Odds Ratio Volcano Plot}
\usage{
plot_odds_volcano(
  odds_ratio_df,
  x = "odds.ratio",
  y = "p.value",
  color = "p.adj",
  labels = "term",
  pCutoff = 0.05
)
}
\arguments{
\item{odds_ratio_df}{Data frame with odds ratio data}

\item{x}{Column name for odds ratio}

\item{y}{Column name for p-value}

\item{color}{Column name for adjusted p-value}

\item{labels}{Column name for labels}

\item{pCutoff}{P-value cutoff for significance}
}
\description{
Function to plot the volcano plot of an odds ratio
}
