% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_percent_genes_detected}
\alias{plot_percent_genes_detected}
\title{Plot Percent Genes Detected}
\usage{
plot_percent_genes_detected(dds, title, min_value = 0)
}
\arguments{
\item{dds}{DESeq2 object}

\item{title}{Title of plot}

\item{min_value}{Minimum value for detection (default: 0)}
}
\value{
ggplot object of percent of genes detected histogram
}
\description{
Function to plot percent of genes detected in each sample
}
