% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_stratified_forest}
\alias{plot_stratified_forest}
\title{Plot Stratified Forest Plot}
\usage{
plot_stratified_forest(
  df,
  x = "x",
  y = "y",
  estimate = "hazard_ratio",
  error_lower = "ci_lower",
  error_upper = "ci_upper",
  color,
  facet
)
}
\arguments{
\item{df}{Data frame to plot}

\item{x}{Column name for x-axis}

\item{y}{Column name for y-axis}

\item{estimate}{Column name for estimates}

\item{error_lower}{Column name for lower error bounds}

\item{error_upper}{Column name for upper error bounds}

\item{color}{Column name for color grouping}

\item{facet}{Column name for facet grouping}
}
\value{
ggplot object of stratified forest plot
}
\description{
Function to create a stratified forest plot from a data frame
}
