% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Plot Volcano Plot}
\usage{
plot_volcano(
  dge,
  x = "log2FoldChange",
  y = "pvalue",
  color = "padj",
  labels = "rownames",
  title = NULL,
  n_labels = TRUE,
  n_gene_labels = 40,
  pCutoff = 0.05,
  xlim = c(min(dge[, x]) - 0.5, max(dge[, x], na.rm = TRUE) + 0.5),
  ylim = c(0, max(-log10(dge[, y]), na.rm = TRUE) * 1.25),
  xlabel = bquote(~Log[2] ~ "Fold Change"),
  ylabel = bquote(~-log[10] ~ "(" ~ .(substitute(pvalue)) ~ ")"),
  color_scale = c("grey", "red")
)
}
\arguments{
\item{dge}{Data frame with differential gene expression data}

\item{x}{Column name for log2 fold change}

\item{y}{Column name for p-value}

\item{color}{Column name for coloring points (if pCutoff not provided, color using this column)}

\item{labels}{Column name for labels}

\item{title}{Title of the plot}

\item{n_labels}{Logical, whether to show number of up/down regulated genes}

\item{pCutoff}{P-value cutoff for significance}

\item{xlim}{Limits for x-axis}

\item{ylim}{Limits for y-axis}
}
\value{
ggplot object
}
\description{
Function to plot the volcano plot
}
