% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_functions.R, R/plotting_functions.R
\name{plotting_functions}
\alias{plotting_functions}
\alias{check_input_data}
\title{Plotting Functions}
\usage{
check_input_data(data, required_columns)
}
\arguments{
\item{data}{Data frame containing the data to be checked}

\item{required_columns}{Vector of required column names}
}
\value{
Logical value indicating whether the data is valid
}
\description{
Functions for pooled cohort risk analysis.

Function to check input data

Functions for plotting data.
}
\author{
Matthew Muller
}
