% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos_functions.R
\name{plotting_track_heatmap}
\alias{plotting_track_heatmap}
\title{Plot Heatmap Track}
\usage{
plotting_track_heatmap(
  circos_data,
  track_name,
  track_height = 0.1,
  col = NULL,
  border = NA
)
}
\arguments{
\item{circos_data}{Prepared circos data object}

\item{track_name}{Name of track to plot}

\item{track_height}{Height of the track (default: 0.1)}

\item{col}{Color function or vector (default: NULL uses blue-white-red)}

\item{border}{Heatmap cell border (default: NA)}
}
\value{
NULL (modifies current circos plot)
}
\description{
Adds a heatmap track to a circos plot.
}
