% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos_functions.R
\name{plotting_track_points}
\alias{plotting_track_points}
\title{Plot Points Track}
\usage{
plotting_track_points(
  circos_data,
  track_name,
  track_height = 0.1,
  col = NULL,
  pch = 16,
  cex = 0.5,
  ylim = "auto",
  bg.col = "#EEEEEE",
  bg.border = NA
)
}
\arguments{
\item{circos_data}{Prepared circos data object}

\item{track_name}{Name of track to plot (from track_cols)}

\item{track_height}{Height of the track (default: 0.1)}

\item{col}{Point color (default: NULL uses circlize default)}

\item{pch}{Point shape (default: 16)}

\item{cex}{Point size (default: 0.5)}

\item{ylim}{Y-axis limits (default: "auto" uses data range)}

\item{bg.col}{Background color (default: "#EEEEEE")}

\item{bg.border}{Background border (default: NA)}
}
\value{
NULL (modifies current circos plot)
}
\description{
Adds a scatter plot track to a circos plot.
}
