% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{press_geneset}
\alias{press_geneset}
\title{Press Geneset}
\format{
A data frame with 100 rows and 3 variables:
\describe{
  \item{gs_name}{Geneset name}
  \item{gene_symbol}{Gene symbol}
}
}
\source{
\url{https://www.nature.com/articles/s41467-024-50994-7}
}
\usage{
press_geneset
}
\description{
A dataframe containing the PRESS geneset
}
\keyword{datasets}
