% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{assay}
\alias{rowData}
\alias{colData}
\alias{select}
\alias{filter}
\alias{mutate}
\alias{summarize}
\alias{arrange}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{arrange}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{select}}, \code{\link[dplyr:summarise]{summarize}}}

  \item{SummarizedExperiment}{\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{colData}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{rowData}}}
}}

