% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{remove_na_variables}
\alias{remove_na_variables}
\title{Function to remove NA variables from a summarized experiment object}
\usage{
remove_na_variables(se, columns)
}
\arguments{
\item{se}{SummarizedExperiment object}

\item{columns}{list of columns to remove NAs from}
}
\value{
DESeqDataSet object with NAs removed
}
\description{
Function to remove NA variables from a summarized experiment object
}
