% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering_functions.R
\name{rna_preprocessing}
\alias{rna_preprocessing}
\title{Run general preprocessing on a DESeq2 object}
\usage{
rna_preprocessing(
  dds,
  outpath,
  depth_filter = 10,
  size_filter = 1e+06,
  percent_filter = 0.5,
  normalize = "none",
  group = NULL
)
}
\arguments{
\item{dds}{DESeq2 object}

\item{outpath}{Path to output directory}

\item{depth_filter}{Minimum library depth to keep}

\item{size_filter}{Minimum library size to keep}

\item{percent_filter}{Minimum percent of genes detected to keep}

\item{normalize}{Normalization method}

\item{group}{Column of interest}
}
\value{
DESeq2 object after preprocessing
}
\description{
Run general preprocessing on a DESeq2 object
}
