% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_functions.R
\name{run_correlations}
\alias{run_correlations}
\title{Run Correlations}
\usage{
run_correlations(
  dds,
  condition,
  normalize = "mor",
  cor.method = "spearman",
  cor.alternative = "two.sided",
  cor.exact = NULL
)
}
\arguments{
\item{dds}{DESeq2 object}

\item{condition}{Column of interest}

\item{normalize}{Normalization method to use}

\item{cor.method}{Correlation method to use}

\item{cor.alternative}{Alternative hypothesis for correlation test}

\item{cor.exact}{Logical indicating whether to use exact correlation test}
}
\value{
Correlation results for each gene
}
\description{
Run Correlations
}
