% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_functions.R
\name{run_deseq}
\alias{run_deseq}
\title{Run Simple Deseq analysis}
\usage{
run_deseq(
  dds,
  outpath,
  contrast = NA,
  ...,
  pvalue = "padj",
  pCutoff = 0.05,
  FCcutoff = 0
)
}
\arguments{
\item{dds}{DESeq2 object}

\item{outpath}{Output path for results}

\item{contrast}{Contrast to run}

\item{...}{Additional arguments to pass to DESeq2}

\item{pvalue}{p-value column name}

\item{pCutoff}{p-value cutoff for volcano plot}

\item{FCcutoff}{Fold change cutoff for volcano plot}
}
\value{
Results of differential expression analysis
}
\description{
Run Simple Deseq analysis
}
