% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_functions.R
\name{run_limma}
\alias{run_limma}
\title{Function to run limma voom on a summarized experiment object with TMM normalization}
\usage{
run_limma(
  se,
  outpath,
  condition,
  controls = NULL,
  pCutoff = 0.05,
  fcCutoff = 0.5,
  topTable.coef = 2,
  topTable.n = Inf,
  topTable.adjust.method = "BH",
  voom.plot = TRUE
)
}
\arguments{
\item{se}{SummarizedExperiment object}

\item{outpath}{Output path for results}

\item{condition}{Vector of conditions}

\item{controls}{Vector of control groups}

\item{pCutoff}{p-value cutoff for volcano plot}

\item{fcCutoff}{Fold change cutoff for volcano plot}

\item{topTable.coef}{Coefficient to test in topTable}

\item{topTable.n}{Number of top results to return}

\item{topTable.adjust.method}{Method for adjusting p-values in topTable}

\item{voom.plot}{Logical indicating whether to plot the voom results}
}
\value{
Data frame of results from limma analysis
}
\description{
Function to run limma on a summarized experiment object with TMM normalization
}
\note{
Requires limma and edgeR packages. Install with: BiocManager::install(c("limma", "edgeR"))
}
