% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_functions.R
\name{save_gse}
\alias{save_gse}
\title{Save and plot gse object}
\usage{
save_gse(gse, outpath, terms2plot = c("inflam", "plat", "coag"))
}
\arguments{
\item{gse}{GSE object.}

\item{outpath}{Path to save to.}

\item{terms2plot}{Terms to plot.}
}
\value{
None.
}
\description{
Save and plot gse object
}
\note{
Requires enrichplot package for plotting. Install with: BiocManager::install("enrichplot")
}
