% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{save_se}
\alias{save_se}
\title{Function to save a SummarizedExperiment object into multiple files if the slots are filled in the object}
\usage{
save_se(se, path, normalize = "mor")
}
\arguments{
\item{se}{SummarizedExperiment object}

\item{path}{path to save files}

\item{normalize}{how to normalize the counts}
}
\description{
Function to save a SummarizedExperiment object into multiple files if the slots are filled in the object
}
