% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{search_vector}
\alias{search_vector}
\title{Search a vector for a string}
\usage{
search_vector(vector, string, ...)
}
\arguments{
\item{vector}{vector, vector to search}

\item{string}{character, string to search for}

\item{...}{additional arguments passed to grep}
}
\value{
named vector of indices
}
\description{
Search a vector for a string
}
