% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_functions.R
\name{select_top_expression}
\alias{select_top_expression}
\title{Filter genes by expression}
\usage{
select_top_expression(df, min_expr = 1)
}
\arguments{
\item{df}{data frame [samples x genes]}

\item{min_expr}{minimum expression value}
}
\value{
data frame with filtered genes [genes, expression]
}
\description{
Filter genes by expression
}
