% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_functions.R
\name{select_top_lasso}
\alias{select_top_lasso}
\title{Filter genes by lasso regression}
\usage{
select_top_lasso(df, y, lambda = NULL, nfolds = 10, ...)
}
\arguments{
\item{df}{data frame [samples x genes]}

\item{y}{response variable}

\item{lambda}{lambda value for lasso regression}

\item{nfolds}{number of cross-validation folds}

\item{...}{additional arguments to pass to lasso function}
}
\value{
data frame with filtered genes [genes, coefficient]
}
\description{
Filter genes by lasso regression
}
