% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_seed}
\alias{set_seed}
\title{Set the seed for reproducibility}
\usage{
set_seed(seed)
}
\arguments{
\item{seed}{integer, seed to set}
}
\value{
seed integer, seed that was set
}
\description{
Set the seed for reproducibility
}
