% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_output_dir}
\alias{setup_output_dir}
\title{Set up a directory for output}
\usage{
setup_output_dir(outpath, overwrite = FALSE)
}
\arguments{
\item{outpath}{path to output directory}

\item{overwrite}{logical, whether to overwrite the directory if it already exists}
}
\value{
outpath path to output directory
}
\description{
Set up a directory for output
}
