% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_functions.R
\name{signature_functions
Select top genes by variance}
\alias{signature_functions
Select top genes by variance}
\alias{select_top_variability}
\title{Signature Functions}
\usage{
select_top_variability(df, n = 1000)
}
\arguments{
\item{df}{data frame [samples x genes]}

\item{n}{number of top genes to select}
}
\value{
data frame with columns [genes, variance]
}
\description{
This script contains functions for signature analysis.

Select top genes by variance
}
