% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_functions.R
\name{signature_functions}
\alias{signature_functions}
\title{Signature Functions}
\description{
This script contains functions for signature analysis.
}
\details{
The functions include selecting top genes by variance, filtering genes by expression, filtering genes by lasso regression, and aligning a signature by average correlation with the derivation values.
}
