% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{stats_functions}
\alias{stats_functions}
\title{stats_functions}
\description{
This script contains functions for statistical analysis.
}
\details{
The functions include creating a statistics table, hypergeometric scoring, creating a correlation matrix, softmaxing a vector, and min-max normalizing a vector.
}
