% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{stats_table}
\alias{stats_table}
\title{Create a Table 1}
\usage{
stats_table(
  data,
  groups,
  vars = NULL,
  printArgs = list(showAllLevels = FALSE, printToggle = FALSE),
  ...
)
}
\arguments{
\item{data}{data.frame, data to make table 1 from}

\item{groups}{character, groups to stratify by}

\item{vars}{character, variables to include in table 1}

\item{printArgs}{list, arguments to pass to print}

\item{...}{other arguments to pass to CreateTableOne}
}
\value{
data.frame, table 1
}
\description{
Create a stats table from a data.frame
}
