% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_functions.R
\name{stratified_ora}
\alias{stratified_ora}
\title{Perform up and down overrepresentation analysis}
\usage{
stratified_ora(
  gene_dataframe,
  outpath,
  method = "enrichGO",
  padj_cutoff = 0.05,
  max_pathways = 5,
  ...
)
}
\arguments{
\item{gene_dataframe}{Data frame with features and direction.}

\item{outpath}{Path to save results.}

\item{method}{Enrichment method (default is "enrichGO").}

\item{padj_cutoff}{Adjusted p-value cutoff (default is 0.05).}

\item{max_pathways}{Maximum number of pathways to display (default is 5).}

\item{...}{Additional arguments to pass to the enrichment function.}
}
\value{
Overrepresentation analysis results.
}
\description{
Perform up and down overrepresentation analysis
}
