% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_functions.R
\name{stratified_overrepresentation}
\alias{stratified_overrepresentation}
\title{Perform up and down overrepresentation analysis using enrichGO}
\usage{
stratified_overrepresentation(
  gene_dataframe,
  outpath,
  padj_cutoff = 0.05,
  max_pathways = 10,
  ont = "BP",
  OrgDb = org.Hs.eg.db::org.Hs.eg.db,
  keyType = "SYMBOL",
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.1,
  minGSSize = 10,
  maxGSSize = 500
)
}
\arguments{
\item{gene_dataframe}{Data frame with 'features' (gene symbols) and 'direction' (up/down) columns.}

\item{outpath}{Path to save results.}

\item{padj_cutoff}{Adjusted p-value cutoff for filtering results (default is 0.05).}

\item{max_pathways}{Maximum number of pathways to display per direction (default is 10).}

\item{ont}{GO ontology: "BP" (Biological Process), "MF" (Molecular Function), "CC" (Cellular Component), or "ALL" (default is "BP").}

\item{OrgDb}{OrgDb object for annotation (default is org.Hs.eg.db).}

\item{keyType}{Key type for gene IDs (default is "SYMBOL").}

\item{pvalueCutoff}{P-value cutoff for enrichGO (default is 0.05).}

\item{qvalueCutoff}{Q-value cutoff for enrichGO (default is 0.1).}

\item{minGSSize}{Minimum gene set size (default is 10).}

\item{maxGSSize}{Maximum gene set size (default is 500).}
}
\value{
List containing enrichment results for up and down genes, combined results, and plot.
}
\description{
Perform up and down overrepresentation analysis using enrichGO
}
