% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{summarize_df}
\alias{summarize_df}
\title{Function to summarize a dataframe}
\usage{
summarize_df(df)
}
\arguments{
\item{df}{dataframe to summarize}
}
\value{
dataframe with summary statistics
}
\description{
Function to summarize a dataframe
}
