% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_functions.R
\name{test_dds}
\alias{test_dds}
\title{Function to run statistical tests on a DESeq2 object using rstatix functions}
\usage{
test_dds(formula, dds, rstatix_test = rstatix::wilcox_test, ...)
}
\arguments{
\item{formula}{Formula to use for the test.}

\item{dds}{DESeq2 object.}

\item{rstatix_test}{rstatix test to use.}

\item{...}{Additional arguments to pass to rstatix_test.}
}
\value{
Data frame of p-values and adjusted p-values.
}
\description{
Function to run statistical tests on a DESeq2 object using rstatix functions
}
