% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{theme_matt}
\alias{theme_matt}
\title{Custom ggplot2 Theme}
\usage{
theme_matt(base_size = 16, base_family = "", ...)
}
\arguments{
\item{base_size}{Base font size}

\item{base_family}{Base font family}

\item{...}{Additional arguments to pass to theme}
}
\value{
ggplot2 theme object
}
\description{
Custom ggplot2 theme based on theme_classic
}
