% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{time_function}
\alias{time_function}
\title{Time a function}
\usage{
time_function(func)
}
\arguments{
\item{func}{function to time}
}
\value{
time time it took to run the function
}
\description{
Time a function
}
