% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{varName}
\alias{varName}
\title{Function to get the variable name}
\usage{
varName(var)
}
\arguments{
\item{var}{variable name}
}
\value{
character, variable name
}
\description{
Function to get the variable name
}
