% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_helpers.R
\name{check_bioc_package}
\alias{check_bioc_package}
\title{Check and load Bioconductor packages}
\usage{
check_bioc_package(pkg)
}
\arguments{
\item{pkg}{Bioconductor package name}
}
\value{
TRUE if package is available, FALSE otherwise
}
\description{
Check and load Bioconductor packages
}
\keyword{internal}
