% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_helpers.R
\name{check_suggested_package}
\alias{check_suggested_package}
\title{Check and load suggested packages}
\usage{
check_suggested_package(pkg, suggestion = NULL)
}
\arguments{
\item{pkg}{Package name to check and load}

\item{suggestion}{Message to show if package is not available}
}
\value{
TRUE if package is available, FALSE otherwise
}
\description{
Check and load suggested packages
}
\keyword{internal}
