% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{color_mapping}
\alias{color_mapping}
\title{Color Mapping}
\usage{
color_mapping(data, custom_colors = NULL, alpha = 1)
}
\arguments{
\item{data}{Data frame containing variables to be mapped}

\item{custom_colors}{Optional list of custom color mappings}

\item{alpha}{Color transparency (0-1, default: 0.8)}
}
\value{
List of color mappings for continuous and categorical variables
}
\description{
Creates color maps for continuous and categorical variables
}
