% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_functions.R
\name{compare_one_to_many}
\alias{compare_one_to_many}
\title{Compare one column to many columns}
\usage{
compare_one_to_many(
  data,
  one,
  many,
  outdir,
  plot = TRUE,
  method.continuous = "spearman",
  method.two_group = "t.test",
  method.multi_group = "anova",
  cor_args = list(),
  t_test_args = list(),
  anova_args = list(),
  wilcox_args = list(),
  kruskal_args = list()
)
}
\arguments{
\item{data}{data frame}

\item{outdir}{output directory}

\item{plot}{logical, whether to plot the results}

\item{method.continuous}{correlation method (default is "spearman")}

\item{method.two_group}{test method for two-group comparisons (default is "t.test")}

\item{method.multi_group}{test method for multi-group comparisons (default is "anova")}

\item{cor_args}{additional arguments to pass to correlation test}

\item{t_test_args}{additional arguments to pass to t-test}

\item{anova_args}{additional arguments to pass to ANOVA test}

\item{wilcox_args}{additional arguments to pass to wilcoxon test}

\item{kruskal_args}{additional arguments to pass to kruskal-wallis test}

\item{col}{column to compare}

\item{cols}{columns to compare to}
}
\value{
list of stats results and plots
}
\description{
Compare one column to many columns
}
