% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_functions.R
\name{deseq_analysis}
\alias{deseq_analysis}
\title{Function to run DESeq2 analysis on a variety of conditions}
\usage{
deseq_analysis(dds, conditions, controls = NULL, outpath)
}
\arguments{
\item{dds}{DESeq2 object}

\item{conditions}{List of conditions to run DESeq2 on}

\item{controls}{List of controls to run DESeq2 on}

\item{outpath}{Path to save results}
}
\value{
List of results of differential expression analysis
}
\description{
Function to run DESeq2 analysis on a variety of conditions with optional controls
}
