% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_functions.R
\name{eigen_ica}
\alias{eigen_ica}
\title{Calculate eigengenes by independent component analysis (WIP)}
\usage{
eigen_ica(
  df,
  outdir,
  n.comp = 1,
  align = FALSE,
  alg.typ = "parallel",
  fun = "logcosh",
  alpha = 1,
  method = "R",
  row.norm = FALSE,
  maxit = 200,
  tol = 1e-04,
  verbose = FALSE
)
}
\arguments{
\item{df}{data frame [samples x genes]}

\item{outdir}{output directory}

\item{n.comp}{number of components to return}

\item{align}{logical, align eigengenes by average expression}

\item{alg.typ}{algorithm type for ICA}

\item{fun}{contrast function for ICA}

\item{alpha}{parameter for contrast function}

\item{method}{method for ICA}

\item{row.norm}{logical, whether to normalize rows}

\item{maxit}{maximum number of iterations}

\item{tol}{tolerance for convergence}

\item{verbose}{logical, print verbose output}
}
\value{
dataframe with eigengenes
}
\description{
Calculate eigengenes by independent component analysis
}
