% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering_functions.R
\name{filter_edgeR}
\alias{filter_edgeR}
\title{Filter a DESeq2 object by expression and make plots of the filtering}
\usage{
filter_edgeR(dds, outpath, group = NULL, min.count = 10, min.prop = 0.5, ...)
}
\arguments{
\item{dds}{DESeq2 object}

\item{outpath}{Path to output directory}

\item{group}{Column of interest for filtering and PCA grouping (optional)}

\item{min.count}{Minimum mean expression to keep (default: 10)}

\item{min.prop}{Minimum proportion of samples to keep (default: 0.5)}

\item{...}{Additional arguments to pass to edgeR::filterByExpr}
}
\value{
DESeq2 object after filtering
}
\description{
Filter genes using edgeR's filterByExpr and generate diagnostic plots
}
