% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_functions.R
\name{get_fc_list}
\alias{get_fc_list}
\title{Get the fold change list}
\usage{
get_fc_list(res, fc_col = "log2FoldChange", names = NULL)
}
\arguments{
\item{res}{Data frame of results.}

\item{fc_col}{Column of fold change values.}

\item{names}{Column of names (if NULL, defaults to rownames).}
}
\value{
Vector of sorted fold change values.
}
\description{
Get the fold change list
}
