% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{hazard_ratios_table}
\alias{hazard_ratios_table}
\title{Hazard Ratios Table}
\usage{
hazard_ratios_table(
  df,
  condition,
  censors,
  controls = NULL,
  per_sd = FALSE,
  ovr = FALSE,
  censor_prefix = "censor_",
  time_prefix = "time_to_"
)
}
\arguments{
\item{df}{data.frame, data to make survival curves from.}

\item{condition}{character, condition to make survival curves for.}

\item{censors}{character, censors to make survival curves for.}

\item{controls}{character, controls to include in the model. Default is NULL.}

\item{per_sd}{logical, whether to standardize the condition. Default is FALSE.}

\item{ovr}{logical, whether to do one vs rest. Default is FALSE.}

\item{censor_prefix}{character, prefix for censor columns. Default is 'C_'.}

\item{time_prefix}{character, prefix for time columns. Default is 'T_'.}
}
\value{
data.frame, hazard ratios table.
}
\description{
Make HR tables for a condition and censors.
}
