% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{hazards.internal}
\alias{hazards.internal}
\title{Internal Hazard Ratios Helper}
\usage{
hazards.internal(cs, data, condition, controls)
}
\arguments{
\item{cs}{list, list with censor and time variable names.}

\item{data}{data.frame, data to make survival curves from.}

\item{condition}{character, a single condition to fit the model for.}

\item{controls}{character, controls to include in the model. Default is NULL.}
}
\value{
data.frame, hazard ratios for the condition.
}
\description{
Internal function to calculate hazard ratios for a single condition.
}
\keyword{internal}
