% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{hazards.internal}
\alias{hazards.internal}
\title{Internal Hazard Ratios Helper}
\usage{
hazards.internal(cs, data, condition, controls)
}
\arguments{
\item{cs}{list, list of censors and time variables.}

\item{data}{data.frame, data to make survival curves from.}

\item{condition}{character, condition(s) to make survival curves for. Can be single or multiple conditions.}

\item{controls}{character, controls to include in the model. Default is NULL.}
}
\value{
data.frame, hazard ratios for the condition(s).
}
\description{
Internal function to calculate hazard ratios for multiple conditions using univariate analyses.
}
\keyword{internal}
