% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenomics_functions.R
\name{make_phenotypes}
\alias{make_phenotypes}
\title{Make Phenotypes from Composite Encoding}
\usage{
make_phenotypes(hesin, encoding)
}
\arguments{
\item{hesin}{Data frame with columns [dnx_hesin_id, Participant ID, field, value]}

\item{encoding}{Data frame with columns [key, field, value]}
}
\value{
Data frame with matched phenotypes
}
\description{
Make Phenotypes from Composite Encoding
}
