% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converting_functions.R
\name{normalize_counts}
\alias{normalize_counts}
\title{Function to normalize dds object and return counts}
\usage{
normalize_counts(dds, method = "mor", log2 = FALSE)
}
\arguments{
\item{dds}{DESeq2 object}

\item{method}{normalization method}

\item{log2}{logical, whether to log2 transform the counts}
}
\value{
normalized counts
}
\description{
Function to normalize DESeq2 dds object and return counts.
}
