% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_functions.R
\name{ovr_deseq_results}
\alias{ovr_deseq_results}
\title{OVR Deseq Results Function}
\usage{
ovr_deseq_results(dds, column, outpath, controls = NULL)
}
\arguments{
\item{dds}{DESeq2 object}

\item{column}{Column of interest for OVR analysis}

\item{outpath}{Output path for results}

\item{controls}{Control groups}

\item{...}{Additional arguments to pass to DESeq2}
}
\value{
OVR results for each level of column of interest
}
\description{
Function to run OVR differential expression analysis on a DESeq2 object
}
