% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering_functions.R
\name{percentGenesDetected}
\alias{percentGenesDetected}
\title{Get percent of genes detected}
\usage{
percentGenesDetected(dds, min_value = 0)
}
\arguments{
\item{dds}{DESeq2 or SummarizedExperiment object}

\item{min_value}{Minimum value to consider a gene detected (default: 0)}
}
\value{
Numeric vector of percent of genes detected per gene
}
\description{
Calculate the proportion of samples in which each gene is detected
}
\keyword{internal}
