% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_functions.R
\name{ReadFeatureCounts}
\alias{ReadFeatureCounts}
\title{Read Feature Counts}
\usage{
ReadFeatureCounts(
  f,
  idx,
  sep = "\\t",
  skip = 1,
  comment.char = "#",
  stringsAsFactors = FALSE
)
}
\arguments{
\item{f}{File path to the feature counts output.}

\item{idx}{Columns to extract for (1) gene name and (2) counts.}

\item{sep}{Field separator character.}

\item{skip}{Number of lines to skip at the beginning of the file.}

\item{comment.char}{Character indicating comment lines to ignore.}

\item{stringsAsFactors}{Logical indicating whether strings should be converted to factors.}
}
\value{
Data frame of counts.
}
\description{
Function to read a feature counts output file.
}
