% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{SummarizedExperiment}
\alias{assay}
\alias{rowData}
\alias{colData}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{SummarizedExperiment}{\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{colData}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{rowData}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}}
}}

