% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_functions.R
\name{run_deseq}
\alias{run_deseq}
\title{Run Simple Deseq analysis}
\usage{
run_deseq(
  dds,
  outpath,
  contrast = NA,
  pvalue = "padj",
  pCutoff = 0.05,
  fcCutoff = 0,
  ...
)
}
\arguments{
\item{dds}{DESeq2 object}

\item{outpath}{Output path for results}

\item{contrast}{Contrast to run}

\item{pvalue}{p-value column name}

\item{pCutoff}{p-value cutoff for volcano plot}

\item{fcCutoff}{Fold change cutoff for volcano plot}

\item{...}{Additional arguments to pass to DESeq2}
}
\value{
Results of differential expression analysis
}
\description{
Function to run a simple DESeq analysis on a summarized experiment object
}
\note{
Requires DESeq2 package. Install with: BiocManager::install("DESeq2")
}
