% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_functions.R
\name{run_limma}
\alias{run_limma}
\title{Function to run limma voom on a summarized experiment object with TMM normalization}
\usage{
run_limma(
  se,
  outpath,
  condition,
  controls = NULL,
  pCutoff = 0.05,
  FCcutoff = 0.5,
  ...
)
}
\arguments{
\item{se}{SummarizedExperiment object}

\item{outpath}{Output path for results}

\item{condition}{Vector of conditions}

\item{controls}{Vector of control groups}

\item{pCutoff}{p-value cutoff for volcano plot}

\item{FCcutoff}{Fold change cutoff for volcano plot}

\item{...}{Additional arguments to pass to voom}
}
\value{
Data frame of p-values and adjusted p-values
}
\description{
Function to run limma on a summarized experiment object with TMM normalization
}
