% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_functions.R
\name{stratified_ora}
\alias{stratified_ora}
\title{Perform up and down overrepresentation analysis}
\usage{
stratified_ora(
  gene_dataframe,
  outpath,
  method = "enrichR",
  dbs = c("GO_Biological_Process_2023", "GO_Cellular_Component_2023",
    "GO_Molecular_Function_2023", "WikiPathway_2023_Human", "GWAS_Catalog_2023",
    "Reactome_2022", "MSigDB Hallmark 2020"),
  padj_cutoff = 0.05,
  max_pathways = 5,
  ...
)
}
\arguments{
\item{gene_dataframe}{Data frame with features and direction columns.}

\item{outpath}{Path to save results.}

\item{method}{Enrichment method: "enrichGO", "groupGO", or "enrichR" (default).}

\item{dbs}{List of databases to use for enrichR (ignored for other methods).}

\item{padj_cutoff}{Adjusted p-value cutoff (default is 0.05).}

\item{max_pathways}{Maximum number of pathways to display (default is 5).}

\item{...}{Additional arguments to pass to the enrichment function.}
}
\value{
Overrepresentation analysis results.
}
\description{
Perform up and down overrepresentation analysis
}
\note{
Requires clusterProfiler, enrichR, or org.Hs.eg.db depending on method. Install with: BiocManager::install(c("clusterProfiler", "enrichR", "org.Hs.eg.db"))
}
