% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{survival_analysis}
\alias{survival_analysis}
\title{Survival Analysis}
\usage{
survival_analysis(
  df,
  condition,
  censors,
  outdir,
  image_type = "pdf",
  censor_prefix = "censor_",
  time_prefix = "time_to_"
)
}
\arguments{
\item{df}{data.frame, data to make survival curves from.}

\item{condition}{character, condition to make survival curves for.}

\item{censors}{character, censors to make survival curves for.}

\item{outdir}{character, output directory to save plots to.}

\item{image_type}{character, type of image to save. Default is 'pdf'.}

\item{censor_prefix}{character, prefix for censor columns. Default is 'C_'.}

\item{time_prefix}{character, prefix for time columns. Default is 'T_'.}
}
\value{
list, list of survival plots and HR plots.
}
\description{
Make survival curves and HR plots for a condition and censors.
}
