% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{survival_analysis}
\alias{survival_analysis}
\title{Survival Analysis}
\usage{
survival_analysis(
  df,
  condition,
  censors,
  outpath,
  censor_prefix = "censor_",
  time_prefix = "time_to_"
)
}
\arguments{
\item{df}{data.frame, data to make survival curves from.}

\item{condition}{character, condition(s) to make survival curves for. Can be a single condition or vector of conditions.}

\item{censors}{character, censors to make survival curves for.}

\item{outpath}{character, output directory to save plots to.}

\item{censor_prefix}{character, prefix for censor columns. Default is 'censor_'.}

\item{time_prefix}{character, prefix for time columns. Default is 'time_to_'.}

\item{image_type}{character, type of image to save. Default is 'pdf'.}
}
\value{
list, list of survival plots and HR plots.
}
\description{
Make survival curves and HR plots for one or more conditions and censors using univariate analyses.
}
